Move parse_model to utils
